fix: path-not found for new defined API routes.
Problem : The RunApp.java file was defined in com.ems.controller package which limited the new API routes to be defined in that package only.

Solution:
Move the RunApp.java file in com.ems package which makes it accessible for the application to reach all the routes defined in the sub-packages.